# backend/internal/transport/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Outfoxe Backend API
  version: 1.0.0
  description: |
    HTTP and WebSocket contract for the Outfoxe backend.

    Authentication for protected HTTP endpoints uses `Authorization: Bearer <access_token>`.

    WebSocket connection:
    - URL: `ws://localhost:8080/ws/games/{id}?token=<access_token>`
    - header alternative: `Authorization: Bearer <access_token>`
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/games:
    post:
      tags: [Games]
      summary: Create game
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGameResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User is already in another unfinished game
          content:
            text/plain:
              schema:
                type: string
                example: user is already in another game
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/games/{id}:
    get:
      tags: [Games]
      summary: Get lobby snapshot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GameID'
      responses:
        '200':
          description: Lobby snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LobbySnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/GameNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /ws/games/{id}:
    get:
      tags: [WebSocket]
      summary: Game WebSocket handshake
      description: |
        Upgrades the HTTP request to a WebSocket connection.

        Authentication:
        - query param `token`, or
        - header `Authorization: Bearer <access_token>`

        After successful connect, server sends an initial `update` message with full state.

        Client message format:
        ```json
        {
          "id": "req-1",
          "type": "command",
          "command": "choose_goal",
          "payload": { "goal": "clue" }
        }
        ```

        Server update message:
        ```json
        {
          "id": "req-1",
          "type": "update",
          "payload": {
            "state": { },
            "events": [ ]
          }
        }
        ```

        Server error message:
        ```json
        {
          "id": "req-1",
          "type": "error",
          "payload": {
            "code": "not_your_turn",
            "message": "It is not your turn."
          }
        }
        ```
      parameters:
        - $ref: '#/components/parameters/GameID'
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: Access token for WebSocket authorization.
      responses:
        '101':
          description: Switching Protocols
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Access denied after upgrade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WSResponseError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GameID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Missing or invalid access token
      content:
        text/plain:
          schema:
            type: string
            example: missing authorization header
    Forbidden:
      description: User does not have access to this game
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
    GameNotFound:
      description: Game not found
      content:
        text/plain:
          schema:
            type: string
            example: game not found
    InternalError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: internal server error
  schemas:
...
    CreateGameRequest:
      type: object
      properties:
        mode:
          type: string
          example: online
...
    GameSummary:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [waiting, active, finished]
    PlayerSummary:
      type: object
      required: [user_id, seat]
      properties:
        user_id:
          type: string
          format: uuid
        seat:
          type: integer
    CreateGameResult:
      type: object
      required: [game, player]
      properties:
        game:
          $ref: '#/components/schemas/GameSummary'
        player:
          $ref: '#/components/schemas/PlayerSummary'
...
    LobbyPlayer:
      type: object
      required: [user_id, seat, display_name, is_me]
      properties:
        user_id:
          type: string
          format: uuid
        seat:
          type: integer
        display_name:
          type: string
        is_me:
          type: boolean
    LobbyGame:
      type: object
      required: [id, status, players, can_start, min_players, max_players]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [waiting, active, finished]
        players:
          type: array
          items:
            $ref: '#/components/schemas/LobbyPlayer'
        can_start:
          type: boolean
        min_players:
          type: integer
          example: 2
        max_players:
          type: integer
          example: 4
    LobbySnapshot:
      type: object
      required: [game]
      properties:
        game:
          $ref: '#/components/schemas/LobbyGame'
...
    WSError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    WSResponseError:
      type: object
      required: [type, payload]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [error]
        payload:
          $ref: '#/components/schemas/WSError'